perf(jsruntime)!: use the native stack instead of Vec<T> (#151)
* perf(jsruntime)!: use the native stack instead of Vec<T>

This change drastically improves the execution time of `fib(41)`.  After the change, `jstb` is 6x
faster than before.  Now, `jstb` is 3x slower than V8 in the case of `fib(41)`.

All arguments and local variables including immutable variables are placed on the native stack.  A
variable is represented with a *tagged* structure and its size is 16 bytes.  So, the native stack
is consumed faster than C/C++.  This may be improved by type inference.  In the future, we might
have to place them in the heap if they consumes large amount of memory in the native stack in order
to avoid a stack overflow.

* refactor(jsruntime): remove unused types and functions

* refactor(jsruntime): optimize the size of generated code

* chore(jsruntime): remove cbindgen.toml

* refactor(jsruntime): refactoring

* perf(jsruntime): allocate variables in the entry block

* fix(jsruntime): disable warning

* feat(jsruntime)!: change the argument type

* feat(jstb): add `compile` sub-command

* chore(jstb): add files

* refactor(jsruntime): refactoring

* refactor(jsruntime): refactoring the llvmir module

* chore(jsruntime): fix clippy

* refactor(jsruntime): use bridge::Locator

* perf(jsruntime): reduce the size of `Binding`

* refactor(jsruntime): remove unused variables

* refactor(jsruntime): use bridge::Value

* fix(jsruntime): clear binding flags

* refactor(jsruntime): re-organize files

* refactor(jsruntime): remove unused type

* refactor(jsruntime): rename

* refactor(jsruntime): refactoring

* fix(jsruntime): fix build.rs

* feat(jsruntime): argument in outer scope